% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TangPoemR.R
\name{write_tangpoem}
\alias{write_tangpoem}
\title{Write Chinese Tang Poems}
\usage{
write_tangpoem(n)
}
\arguments{
\item{n, }{should be the length of the poem}
}
\description{
This package write Chinese Tang Poems automatically
}
\examples{
write_tangpoem(5)
}
